7주차 > advanced > UI 기준으로 컴포넌트 및 훅 분리
