+++ packages/assignment-7/src/contants.ts
@@ -1,0 +1,35 @@
+export const categories = ["업무", "개인", "가족", "기타"];
+
+export const weekDays = ["일", "월", "화", "수", "목", "금", "토"];
+
+export const notificationOptions = [
+  { value: 1, label: "1분 전" },
+  { value: 10, label: "10분 전" },
+  { value: 60, label: "1시간 전" },
+  { value: 120, label: "2시간 전" },
+  { value: 1440, label: "1일 전" },
+];
+
+// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+// @ts-ignore
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+export const fetchHolidays = (year: number, month: number) => {
+  // 실제로는 API를 호출하여 공휴일 정보를 가져와야 합니다.
+  // 여기서는 예시로 하드코딩된 데이터를 사용합니다.
+  return {
+    "2024-01-01": "신정",
+    "2024-02-09": "설날",
+    "2024-02-10": "설날",
+    "2024-02-11": "설날",
+    "2024-03-01": "삼일절",
+    "2024-05-05": "어린이날",
+    "2024-06-06": "현충일",
+    "2024-08-15": "광복절",
+    "2024-09-16": "추석",
+    "2024-09-17": "추석",
+    "2024-09-18": "추석",
+    "2024-10-03": "개천절",
+    "2024-10-09": "한글날",
+    "2024-12-25": "크리스마스",
+  };
+};
