+++ packages/assignment-7/src/hooks/useCalendarView.ts
@@ -1,0 +1,41 @@
+import { useEffect, useState } from "react";
+import { fetchHolidays } from "../contants";
+import { ViewType } from "../types";
+
+const useCalendarView = () => {
+  const [view, setView] = useState<ViewType>("month");
+  const [currentDate, setCurrentDate] = useState(new Date());
+  const [holidays, setHolidays] = useState<{ [key: string]: string }>({});
+
+  /**
+   * 달력 전 후 이동
+   */
+  const navigate = (direction: "prev" | "next") => {
+    setCurrentDate((prevDate) => {
+      const newDate = new Date(prevDate);
+      if (view === "week") {
+        newDate.setDate(newDate.getDate() + (direction === "next" ? 7 : -7));
+      } else if (view === "month") {
+        newDate.setMonth(newDate.getMonth() + (direction === "next" ? 1 : -1));
+      }
+      return newDate;
+    });
+  };
+
+  useEffect(() => {
+    const year = currentDate.getFullYear();
+    const month = currentDate.getMonth() + 1;
+    const newHolidays = fetchHolidays(year, month);
+    setHolidays(newHolidays);
+  }, [currentDate]);
+
+  return {
+    view,
+    setView,
+    currentDate,
+    holidays,
+    navigate,
+  };
+};
+
+export default useCalendarView;
