+++ packages/assignment-7/src/hooks/useEventForm.ts
@@ -1,0 +1,197 @@
+import { useToast } from "@chakra-ui/react";
+import { ChangeEvent, useMemo, useState } from "react";
+import { Event, RepeatType, SetState } from "../types";
+
+interface Props {
+  setEditingEvent: SetState<Event | null>;
+}
+const useEventForm = ({ setEditingEvent }: Props) => {
+  const toast = useToast();
+
+  const [title, setTitle] = useState("");
+  const [date, setDate] = useState("");
+  const [startTime, setStartTime] = useState("");
+  const [endTime, setEndTime] = useState("");
+  const [description, setDescription] = useState("");
+  const [location, setLocation] = useState("");
+  const [category, setCategory] = useState("");
+  const [isRepeating, setIsRepeating] = useState(false);
+  const [repeatType, setRepeatType] = useState<RepeatType>("none");
+  const [repeatInterval, setRepeatInterval] = useState(1);
+  const [repeatEndDate, setRepeatEndDate] = useState("");
+  const [notificationTime, setNotificationTime] = useState(10);
+
+  const [startTimeError, setStartTimeError] = useState<string | null>(null);
+  const [endTimeError, setEndTimeError] = useState<string | null>(null);
+
+  /**
+   * 시작시간, 종료시간이 관계적으로 성립하는지 확인
+   */
+  const validateTime = (start: string, end: string) => {
+    if (!start || !end) return;
+
+    const startDate = new Date(`2000-01-01T${start}`);
+    const endDate = new Date(`2000-01-01T${end}`);
+
+    if (startDate >= endDate) {
+      setStartTimeError("시작 시간은 종료 시간보다 빨라야 합니다.");
+      setEndTimeError("종료 시간은 시작 시간보다 늦어야 합니다.");
+    } else {
+      setStartTimeError(null);
+      setEndTimeError(null);
+    }
+  };
+
+  /**
+   * 시작시간 변경
+   */
+  const handleStartTimeChange = (e: ChangeEvent<HTMLInputElement>) => {
+    const newStartTime = e.target.value;
+    setStartTime(newStartTime);
+    validateTime(newStartTime, endTime);
+  };
+
+  /**
+   * 종료시간 변경
+   */
+  const handleEndTimeChange = (e: ChangeEvent<HTMLInputElement>) => {
+    const newEndTime = e.target.value;
+    setEndTime(newEndTime);
+    validateTime(startTime, newEndTime);
+  };
+
+  /**
+   * 입력 초기화
+   */
+  const resetForm = () => {
+    setTitle("");
+    setDate("");
+    setStartTime("");
+    setEndTime("");
+    setDescription("");
+    setLocation("");
+    setCategory("");
+    setEditingEvent(null);
+    setIsRepeating(false);
+    setRepeatType("none");
+    setRepeatInterval(1);
+    setRepeatEndDate("");
+  };
+
+  /**
+   * 특정 일정 수정하기
+   */
+  const editEvent = (event: Event) => {
+    setEditingEvent(event);
+    setTitle(event.title);
+    setDate(event.date);
+    setStartTime(event.startTime);
+    setEndTime(event.endTime);
+    setDescription(event.description);
+    setLocation(event.location);
+    setCategory(event.category);
+    setIsRepeating(event.repeat.type !== "none");
+    setRepeatType(event.repeat.type);
+    setRepeatInterval(event.repeat.interval);
+    setRepeatEndDate(event.repeat.endDate || "");
+    setNotificationTime(event.notificationTime);
+  };
+
+  const eventFormData = useMemo<Omit<Event, "id">>(() => {
+    const formData = {
+      title,
+      date,
+      startTime,
+      endTime,
+      description,
+      location,
+      category,
+      repeat: {
+        type: isRepeating ? repeatType : "none",
+        interval: repeatInterval,
+        endDate: repeatEndDate || undefined,
+      },
+      notificationTime,
+    };
+    return { ...formData };
+  }, [
+    title,
+    date,
+    startTime,
+    endTime,
+    description,
+    location,
+    category,
+    isRepeating,
+    notificationTime,
+    repeatType,
+    repeatInterval,
+    repeatEndDate,
+  ]);
+
+  const onBlurTimeInput = () => validateTime(startTime, endTime);
+
+  const validateSubmitEventForm = async () => {
+    if (!title || !date || !startTime || !endTime) {
+      toast({
+        title: "필수 정보를 모두 입력해주세요.",
+        status: "error",
+        duration: 3000,
+        isClosable: true,
+      });
+      return false;
+    }
+
+    validateTime(startTime, endTime);
+    if (startTimeError || endTimeError) {
+      toast({
+        title: "시간 설정을 확인해주세요.",
+        status: "error",
+        duration: 3000,
+        isClosable: true,
+      });
+      return false;
+    }
+
+    return true;
+  };
+
+  return {
+    title,
+    setTitle,
+    date,
+    setDate,
+    startTime,
+    handleStartTimeChange,
+    endTime,
+    handleEndTimeChange,
+    description,
+    setDescription,
+    location,
+    setLocation,
+    category,
+    setCategory,
+    isRepeating,
+    setIsRepeating,
+    notificationTime,
+    setNotificationTime,
+    repeatType,
+    setRepeatType,
+    repeatInterval,
+    setRepeatInterval,
+    repeatEndDate,
+    setRepeatEndDate,
+
+    startTimeError,
+    endTimeError,
+    onBlurTimeInput,
+    validateSubmitEventForm,
+
+    resetForm,
+    editEvent,
+
+    eventFormData,
+  };
+};
+
+export default useEventForm;
