+++ packages/assignment-7/src/types.ts
@@ -1,4 +1,10 @@
+import { Dispatch, SetStateAction } from "react";
+
+export type ViewType = "week" | "month";
+
 export type RepeatType = "none" | "daily" | "weekly" | "monthly" | "yearly";
+
+export type Notification = { id: number; message: string };
 
 export interface RepeatInfo {
   type: RepeatType;
@@ -18,3 +24,5 @@ export interface Event {
   repeat: RepeatInfo;
   notificationTime: number; // 분 단위로 저장
 }
+
+export type SetState<T> = Dispatch<SetStateAction<T>>;
